chore: cleaned up dev config

+++ scripts/bootstrap.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+# TODO Convert to node script
 
 set -e # Exit as soon as any line in the bash script fails
 
@@ -9,30 +10,32 @@ NO_COLOR="\033[0m" # No color
 
 cd $(dirname "$0")/..
 
-# Remove all yarn locks
-find . -type f -name 'yarn.lock' -not -path "*/node_modules/*" -exec sh -c 'x="{}"; mv "$x" "${x}_"' \;
-# Remove packages's modules
-find . -type d -name "node_modules" -prune -exec rm -rf {} \;
-# Remove builds
-find . -type d -name "lib" -prune -exec rm -rf {} \;
-find . -type d -name "dist" -prune -exec rm -rf {} \;
-find . -type d -name "build" -prune -exec rm -rf {} \;
+if [ ! -d "node_modules" ]; then
+  # Remove all yarn locks
+  find . -type f -name 'yarn.lock' -not -path "*/node_modules/*" -exec sh -c 'x="{}"; mv "$x" "${x}_"' \;
+  # Remove packages's modules
+  find . -type d -name "node_modules" -prune -exec rm -rf {} \;
 
-# Install dependenices
-echo -e "${YELLOW}Installing dependencies...${NO_COLOR}"
-yarn install --silent
+  # Install dependenices
+  echo -e "${YELLOW}Installing dependencies...${NO_COLOR}"
+  yarn install --silent
+
+  # Restore all yarn locks
+  find . -type f -name 'yarn.lock_' -not -path "*/node_modules/*" -exec sh -c 'f="{}"; mv "$f" "${f%_*}"' \;
+
+  echo -e "${YELLOW}Installing dependencies finished${NO_COLOR}"
+else
+  echo -e "${YELLOW}Skipped installing dependencies${NO_COLOR}"
+fi
+
+# Build AdminJS types first
+echo -e "${YELLOW}Building AdminJS example app deps...${NO_COLOR}"
+yarn workspace @adminjs/design-system build
+yarn workspace adminjs types
+node ./scripts/build-example-dependencies.js
 
 # Build all packages
 echo -e "${YELLOW}Building packages...${NO_COLOR}"
-# Build AdminJS types first
-yarn workspace @adminjs/design-system build
-yarn workspace adminjs types
-yarn workspace @adminjs/express build
-yarn workspace @adminjs/hapi build
-yarn workspace @adminjs/mikroorm build
-yarn workspace @adminjs/mongoose build
-yarn workspace @adminjs/prisma build
-yarn workspace @adminjs/typeorm build
 yarn workspaces --no-progress run build
 
 # Build AdminJS
@@ -48,8 +51,5 @@ yarn workspace @adminjs/design-system bundle
 echo -e "${YELLOW}Setting up example app...${NO_COLOR}"
 yarn workspace example docker:up
 yarn workspace example migration:up
-
-# Restore all yarn locks
-find . -type f -name 'yarn.lock_' -not -path "*/node_modules/*" -exec sh -c 'f="{}"; mv "$f" "${f%_*}"' \;
 
 echo -e "${GREEN}Success! All AdminJS bootstrapped.${NO_COLOR}"
